+++ spec/mock_ace.lua
@@ -324,6 +324,14 @@ function MockAce.install()
         widget.frame = {
             IsShown = function() return widget._shown end,
             SetClampedToScreen = function() end,
+            CreateFontString = function()
+                local fs = { _text = "" }
+                fs.SetFont = function() end
+                fs.SetPoint = function() end
+                fs.SetText = function(_, t) fs._text = t end
+                fs.GetText = function() return fs._text end
+                return fs
+            end,
         }
         widget.SetDisabled = function(self, d) self._disabled = d end
         widget.SetLayout = function() end